+++ src/main/java/cz/osu/weaponeshop/service/UserServiceImpl.java
@@ -1,14 +1,13 @@
 package cz.osu.weaponeshop.service;
 
 import cz.osu.weaponeshop.config.JwtService;
-import cz.osu.weaponeshop.exception.ForbiddenException;
-import cz.osu.weaponeshop.exception.UserAlreadyExistsException;
+import cz.osu.weaponeshop.exception.BadRequestException;
 import cz.osu.weaponeshop.exception.NotFoundException;
 import cz.osu.weaponeshop.model.Role;
 import cz.osu.weaponeshop.model.User;
-import cz.osu.weaponeshop.model.dto.LoginRequest;
-import cz.osu.weaponeshop.model.dto.RegisterRequest;
-import cz.osu.weaponeshop.model.dto.UpdateRequest;
+import cz.osu.weaponeshop.model.dto.user.LoginRequest;
+import cz.osu.weaponeshop.model.dto.user.RegisterRequest;
+import cz.osu.weaponeshop.model.dto.user.UpdateRequest;
 import cz.osu.weaponeshop.model.response.AuthenticationResponse;
 import cz.osu.weaponeshop.repository.UserRepository;
 import de.mkammerer.argon2.Argon2;
@@ -25,26 +24,47 @@ public class UserServiceImpl {
     private final JwtService jwtService;
     private final AuthenticationManager authenticationManager;
     private final Argon2 argon2 = Argon2Factory.create();
+    private final String USER_NOT_FOUND = "User was not found!";
+    private final String USER_ALREADY_EXISTS = "User by that name already exists";
+    private final String USER_OR_PASSWORD_EMPTY = "Username or password cannot be empty!";
+    private final String USER_INVALID_PASSWORD = "Invalid password";
 
     public void updateUser(UpdateRequest userDTO) {
         if (userRepo.existsByUserName(userDTO.getNewUserName())) {
-          throw new UserAlreadyExistsException("User by that name already exists");
+            throw new BadRequestException(USER_ALREADY_EXISTS);
         }
         User user = userRepo.findByUserName(userDTO.getCurrentUserName()).orElseThrow(
-                () -> new NotFoundException("User was not found"));
+                () -> new NotFoundException(USER_NOT_FOUND));
         user.setUserName(userDTO.getNewUserName());
         userRepo.save(user);
     }
 
     public void deleteUser(RegisterRequest userDTO) {
+        if (userDTO.isEmpty()){
+            throw new BadRequestException(USER_OR_PASSWORD_EMPTY);
+        }
         User user = userRepo.findByUserName(userDTO.getUserName()).orElseThrow(
-                () -> new NotFoundException("User was not found"));
+                () -> new NotFoundException(USER_NOT_FOUND));
+        if (!argon2.verify(user.getPassword(), userDTO.getPassword().toCharArray())) {
+            throw new BadRequestException(USER_INVALID_PASSWORD);
+        }
         userRepo.delete(user);
     }
 
+    public String getUser(Long id) {
+        if (id == null){
+            throw new BadRequestException("User id cannot be empty");
+        }
+        User user = userRepo.findById(id).orElseThrow(() -> new NotFoundException(USER_NOT_FOUND));
+        return user.getUsername();
+    }
+
     public AuthenticationResponse register(RegisterRequest request) {
-        if (userRepo.findByUserName(request.getUserName()).isPresent()){
-            throw new UserAlreadyExistsException("User by that name already exists");
+        if (request.isEmpty()) {
+            throw new BadRequestException(USER_OR_PASSWORD_EMPTY);
+        }
+        if (userRepo.findByUserName(request.getUserName()).isPresent()) {
+            throw new BadRequestException(USER_ALREADY_EXISTS);
         }
         var user = User.builder()
                 .userName(request.getUserName())
@@ -59,10 +79,13 @@ public class UserServiceImpl {
     }
 
     public AuthenticationResponse login(LoginRequest request) {
+        if (request.isEmpty()) {
+            throw new BadRequestException(USER_OR_PASSWORD_EMPTY);
+        }
         var user = userRepo.findByUserName(request.getUserName())
-                .orElseThrow();
+                .orElseThrow(() -> new NotFoundException(USER_NOT_FOUND));
         if (!argon2.verify(user.getPassword(), request.getPassword().toCharArray())) {
-            throw new ForbiddenException("Invalid password");
+            throw new BadRequestException(USER_INVALID_PASSWORD);
         }
         authenticationManager.authenticate(
                 new UsernamePasswordAuthenticationToken(
@@ -70,7 +93,6 @@ public class UserServiceImpl {
                         request.getPassword()
                 )
         );
-
         var jwtToken = jwtService.generateToken(user);
         return AuthenticationResponse.builder()
                 .token(jwtToken)
